Multiple changes: * Proper nearest size calculation in PAL. * Make sure to dispose the PAL handle properly * Fix alignment issue at p/invoke of rvn_mmap_file() * Make sure "Validate" compilation mode works * Some precautions and some 'printf' debugging (conditional)

+++ src/Raven.Pal/inc/rvn.h
@@ -1,6 +1,8 @@
 #ifndef RVN_H
 #define RVN_H
 
+//#define PRINTF_DEBUG
+
 #ifdef __APPLE__
 #define EXPORT __attribute__((visibility("default")))
 #define PRIVATE __attribute__((visibility("hidden")))
@@ -107,7 +109,7 @@ EXPORT int32_t
 rvn_remap(void *base_address, void **new_address, void *handle, int64_t size, int32_t flags, int64_t offset, int32_t *detailed_error_code);
 
 EXPORT int32_t
-rvn_mmap_file(int64_t sz, int64_t flags, void *handle, int64_t offset, void **addr, int32_t *detailed_error_code);
+rvn_mmap_file(int64_t sz, int32_t flags, void *handle, int64_t offset, void **addr, int32_t *detailed_error_code);
 
 EXPORT int32_t
 rvn_create_file(const char *path, int64_t initial_file_size, int32_t flags, void **handle, int64_t *actual_file_size, int32_t *detailed_error_code);
